requirements.txt
python-dotenv>=1.0.0
ts3>=2.0.0b2
